+++ src/main/java/com/najasin/domain/answer/entity/AnswerId.java
@@ -1,6 +1,6 @@
 package com.najasin.domain.answer.entity;
 
-import com.najasin.domain.question.entity.Question;
+import com.najasin.domain.myManual.question.entity.Question;
 import com.najasin.domain.user.entity.User;
 import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
