+++ src/main/java/com/najasin/domain/userKeyword/entity/UserKeywordId.java
@@ -1,6 +1,6 @@
 package com.najasin.domain.userKeyword.entity;
 
-import com.najasin.domain.keyword.entity.Keyword;
+import com.najasin.domain.myManual.keyword.entity.Keyword;
 import com.najasin.domain.user.entity.User;
 import jakarta.persistence.Embeddable;
 import lombok.Getter;
